+++ WhoAskedBackend/Services/Messaging/IMessageProvider.cs
@@ -4,9 +4,9 @@ namespace WhoAskedBackend.Services.Messaging
 {
     public interface IMessageProvider
     {
-        void ImportQueue(int queueId);
-        void ExportQueue(int queueId);
-        List<Message>? RetrieveLatestMessages(int queueId, int amount);
+        void ImportQueue(long queueId);
+        void ExportQueue(long queueId);
+        List<Message>? RetrieveLatestMessages(long queueId, int amount);
         void QueueMessage(Message message);
     }
 }
